requirements.txt
anthropic>=0.40.0
matplotlib>=3.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
yfinance>=0.2.40
curl_cffi>=0.5.10
duckduckgo-search>=6.0.0
pypdf>=4.0.0
PyMuPDF>=1.24.0
markdown>=3.5.0
